Backend/requirements.txt
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2